# rag-mcp-server/api/openapi.yaml
openapi: 3.0.3
info:
  title: RAG MCP Server API
  description: |
    REST API for the RAG MCP Server - a semantic document search system
    with embedding-based retrieval, cross-encoder reranking, and LLM-powered
    question answering with optional verification.
  version: 1.0.0
  contact:
    name: RAG MCP Server
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Local development server
...
paths:
...
  /collections:
    get:
      summary: List collections
      description: Retrieve all document collections
      tags:
        - collections
      responses:
        '200':
          description: List of collections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionListResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create collection
      description: Create a new document collection
      tags:
        - collections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
      responses:
        '201':
          description: Collection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '409':
          description: Collection name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{id}:
    get:
      summary: Get collection by ID
      description: Retrieve a single collection by its ID
      tags:
        - collections
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: Collection found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update collection
      description: Update a collection's name, description, or color
      tags:
        - collections
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollectionRequest'
      responses:
        '200':
          description: Collection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Name conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete collection
      description: Delete a collection (documents are unassigned, not deleted)
      tags:
        - collections
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: Collection deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      collectionId:
                        type: string
                      message:
                        type: string
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{id}/documents:
    get:
      summary: Get collection documents
      description: Get all documents in a collection
      tags:
        - collections
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: Collection with documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      collection:
                        $ref: '#/components/schemas/Collection'
                      documents:
                        type: array
                        items:
                          $ref: '#/components/schemas/Document'
                      total:
                        type: integer
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Add document to collection
      description: Add an existing document to a collection
      tags:
        - collections
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - documentId
              properties:
                documentId:
                  type: string
      responses:
        '200':
          description: Document added to collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      collectionId:
                        type: string
                      documentId:
                        type: string
                      message:
                        type: string
  /collections/{id}/documents/{documentId}:
    delete:
      summary: Remove document from collection
      description: Remove a document from a collection (document is not deleted)
      tags:
        - collections
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          description: Document ID
          schema:
            type: string
      responses:
        '200':
          description: Document removed from collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      collectionId:
                        type: string
                      documentId:
                        type: string
                      message:
                        type: string
...
  /analytics/trends:
    get:
      summary: Get query trends
      description: Get query trends over time
      tags:
        - analytics
      parameters:
        - name: days
          in: query
          description: Number of days to include
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 7
      responses:
        '200':
          description: Query trends
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryTrendsResponse'
  /analytics/top-queries:
    get:
      summary: Get top queries
      description: Get most frequent queries
      tags:
        - analytics
      parameters:
        - name: limit
          in: query
          description: Maximum results
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: type
          in: query
          description: Filter by query type
          schema:
            type: string
            enum: [search, ask]
      responses:
        '200':
          description: Top queries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopQueriesResponse'
  /analytics/queries:
    get:
      summary: Get recent queries
      description: Get recent query log entries
      tags:
        - analytics
      parameters:
        - name: limit
          in: query
          description: Maximum results
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: type
          in: query
          description: Filter by query type
          schema:
            type: string
            enum: [search, ask]
      responses:
        '200':
          description: Recent queries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentQueriesResponse'
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        filepath:
          type: string
        fileType:
          type: string
          enum: [txt, md, docx, pdf, pptx, xlsx, csv, html, json, rtf]
        fileSize:
          type: integer
          description: File size in bytes
        mimeType:
          type: string
        status:
          type: string
          enum: [pending, processing, indexed, failed]
        chunkCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        indexedAt:
          type: string
          format: date-time
          nullable: true
        summary:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: Document not found
    Collection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        documentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCollectionRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          default: '#6366f1'
    UpdateCollectionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
    CollectionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Collection'
    CollectionListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            collections:
              type: array
              items:
                $ref: '#/components/schemas/Collection'
            total:
              type: integer
...
    QueryTrend:
      type: object
      properties:
        date:
          type: string
          format: date
        searchCount:
          type: integer
        askCount:
          type: integer
        totalCount:
          type: integer
    QueryTrendsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            trends:
              type: array
              items:
                $ref: '#/components/schemas/QueryTrend'
            period:
              type: object
              properties:
                days:
                  type: integer
                start:
                  type: string
                  format: date-time
                end:
                  type: string
                  format: date-time
    TopQuery:
      type: object
      properties:
        query:
          type: string
        count:
          type: integer
        avgLatencyMs:
          type: integer
        lastUsed:
          type: string
          format: date-time
    TopQueriesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            queries:
              type: array
              items:
                $ref: '#/components/schemas/TopQuery'
            total:
              type: integer
    QueryLogEntry:
      type: object
      properties:
        id:
          type: integer
        query:
          type: string
        queryType:
          type: string
          enum: [search, ask]
        source:
          type: string
        resultCount:
          type: integer
        topScore:
          type: number
          nullable: true
        latencyMs:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    RecentQueriesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            queries:
              type: array
              items:
                $ref: '#/components/schemas/QueryLogEntry'
            total:
              type: integer